extract changes to StorageBusInventoryHandler
